Config vite base and for each package
