+++ packages/playground/main.ts
@@ -1,14 +1,4 @@
 import { createApp } from 'vue'
 import App from './App.vue'
 
-if (!Promise.withResolvers) {
-  Promise.withResolvers = function () {
-    let resolve, reject
-    const promise = new Promise((res, rej) => {
-      resolve = res
-      reject = rej
-    })
-    return { promise, resolve, reject }
-  }
-}
 createApp(App).mount('#app')
