--- packages/playground/src/samples/Pages.vue
@@ -1,27 +1,0 @@
-<script setup lang="ts">
-import { ref } from 'vue';
-import { VuePDF, usePDF } from '../../src';
-import pdfdoc from '../pdf/example_014.pdf';
-
-const { pdf } = usePDF(pdfdoc)
-const page = ref(1)
-</script>
-
-<template>
-  <div>
-    <button @click="page = page + 1">
-      + Page
-    </button>
-    <button @click="page = page - 1">
-      - Page
-    </button>
-
-    <div style="text-align: center;">
-      <VuePDF :pdf="pdf" :page="page" text-layer watermark-text="confidencial" />
-    </div>
-  </div>
-</template>
-
-<style>
-
-</style>
