+++ packages/vue-pdf/package.json
@@ -22,9 +22,9 @@
   ],
   "exports": {
     ".": {
+      "types": "./dist/types/index.d.ts",
       "require": "./dist/index.umd.js",
-      "import": "./dist/index.mjs",
-      "types": "./dist/types/index.d.ts"
+      "import": "./dist/index.mjs"
     },
     "./style.css": "./dist/style.css",
     "./src/*": "./src/*"
@@ -37,15 +37,11 @@
     "src"
   ],
   "scripts": {
-    "dev": "vite --force --config vite.playground.ts",
     "build": "npm run tests && npm run build:lib && npm run build:dts",
     "build:lib": "vite build",
     "build:dts": "vue-tsc --declaration --emitDeclarationOnly -p tsconfig.build.json",
-    "preview": "vite preview --port 5050",
     "lint": "eslint .",
-    "lint:fix": "eslint --fix .",
-    "test": "vitest run",
-    "tests": "npm run test"
+    "lint:fix": "eslint --fix ."
   },
   "peerDependencies": {
     "vue": "^3.2.33"
