+++ packages/vue-pdf/shims.d.ts
@@ -3,8 +3,3 @@ declare module "*.vue" {
   const component: DefineComponent<{}, {}, any>;
   export default component;
 }
-
-declare module '*.pdf' {
-  const pdfurl: string
-  export default pdfurl
-}
