+++ packages/vue-pdf/tsconfig.build.json
@@ -3,7 +3,5 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "outDir": "dist/types"
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
+  }
 }
