+++ packages/vue-pdf/vite.config.ts
@@ -1,33 +1,27 @@
 import { resolve } from 'node:path'
-
-import vue from '@vitejs/plugin-vue'
-import { defineConfig } from 'vite'
+import { defineConfig, mergeConfig } from 'vite'
+import commonConfig from '../../vite.config'
 
 // https://vitejs.dev/config/
-export default defineConfig({
-  optimizeDeps: {
-    esbuildOptions: {
-      supported: {
-        'top-level-await': true,
+export default mergeConfig(
+  commonConfig,
+  defineConfig({
+    build: {
+      lib: {
+        entry: resolve(__dirname, './src/index.ts'),
+        name: '@tato30/vue-pdf',
+        fileName: 'index',
       },
-    },
-  },
-  build: {
-    lib: {
-      entry: resolve(__dirname, './src/index.ts'),
-      name: '@tato30/vue-pdf',
-      fileName: 'index',
-    },
-    rollupOptions: {
-      external: ['vue', 'pdfjs-dist'],
-      output: {
-        exports: 'named',
-        globals: {
-          'vue': 'vue',
-          'pdfjs-dist': 'PDFJS',
+      rollupOptions: {
+        external: ['vue', 'pdfjs-dist'],
+        output: {
+          exports: 'named',
+          globals: {
+            'vue': 'vue',
+            'pdfjs-dist': 'PDFJS',
+          },
         },
       },
     },
-  },
-  plugins: [vue()],
-})
+  }),
+)
