+++ tests/sizing.spec.ts
@@ -1,11 +1,8 @@
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-/* eslint-disable @typescript-eslint/no-unsafe-return */
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
 import { beforeAll, expect, test, vi } from 'vitest'
 
 import { mount } from '@vue/test-utils'
 
-import { VuePDF, usePDF } from '../src'
+import { VuePDF, usePDF } from '@tato30/vue-pdf'
 
 const { pdf } = usePDF(
   'https://raw.githubusercontent.com/mozilla/pdf.js/ba2edeae/web/compressed.tracemonkey-pldi-09.pdf',
